#36473528 Сделал рефач и убрал факторку создания заказа, вынес все это в сервис заказов.
