+++ LeokaEstetica.Platform.Processing/AutofacModules/ProcessingModule.cs
@@ -1,11 +1,9 @@
 using Autofac;
 using LeokaEstetica.Platform.Core.Attributes;
 using LeokaEstetica.Platform.Processing.Abstractions.Commerce;
-using LeokaEstetica.Platform.Processing.Abstractions.PayMaster;
 using LeokaEstetica.Platform.Processing.Abstractions.YandexKassa;
 using LeokaEstetica.Platform.Processing.Builders.Order;
 using LeokaEstetica.Platform.Processing.Services.Commerce;
-using LeokaEstetica.Platform.Processing.Services.PayMaster;
 using LeokaEstetica.Platform.Processing.Services.YandexKassa;
 using LeokaEstetica.Platform.Processing.Strategies.PaymentSystem;
 
@@ -17,12 +15,12 @@ public class ProcessingModule : Module
     public static void InitModules(ContainerBuilder builder)
     {
         // Сервис платежной системы PayMaster.
-        builder.RegisterType<PayMasterService>()
-            .Named<IPayMasterService>("PayMasterService")
-            .InstancePerLifetimeScope();
-        builder.RegisterType<PayMasterService>()
-            .As<IPayMasterService>()
-            .InstancePerLifetimeScope();
+        // builder.RegisterType<PayMasterService>()
+        //     .Named<IPayMasterService>("PayMasterService")
+        //     .InstancePerLifetimeScope();
+        // builder.RegisterType<PayMasterService>()
+        //     .As<IPayMasterService>()
+        //     .InstancePerLifetimeScope();
 
         // Сервис коммерции.
         builder.RegisterType<CommerceService>()
@@ -63,6 +61,13 @@ public class ProcessingModule : Module
             .InstancePerLifetimeScope();
         builder.RegisterType<PostVacancyOrderBuilder>()
             .As<BaseOrderBuilder>()
+            .InstancePerLifetimeScope();
+            
+        builder.RegisterType<OrdersService>()
+            .Named<IOrdersService>("OrdersService")
+            .InstancePerLifetimeScope();
+        builder.RegisterType<OrdersService>()
+            .As<IOrdersService>()
             .InstancePerLifetimeScope();
     }
 }
